+++ maps-sdk/src/main/java/com/mapbox/maps/renderer/MapboxRenderThread.kt
@@ -84,7 +84,8 @@ internal class MapboxRenderThread : Choreographer.FrameCallback {
   @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
   internal var paused = false
 
-  internal var renderThreadRecorder: RenderThreadRecorder? = null
+  @OptIn(MapboxExperimental::class)
+  internal var renderThreadStatsRecorder: RenderThreadStatsRecorder? = null
 
   /**
    * We track moment when native renderer is prepared.
@@ -356,8 +357,9 @@ internal class MapboxRenderThread : Choreographer.FrameCallback {
     GLES20.glBindBuffer(GLES20.GL_ELEMENT_ARRAY_BUFFER, 0)
   }
 
+  @OptIn(MapboxExperimental::class)
   private fun draw(frameTimeNanos: Long) {
-    if (!fpsManager.preRender(frameTimeNanos, renderThreadRecorder?.recording == true)) {
+    if (!fpsManager.preRender(frameTimeNanos, renderThreadStatsRecorder?.isRecording == true)) {
       // when we have FPS limited and desire to skip core render - we must schedule new draw call
       // otherwise map may remain in not fully loaded state
       postPrepareRenderFrame()
@@ -603,10 +605,11 @@ internal class MapboxRenderThread : Choreographer.FrameCallback {
     }
   }
 
+  @OptIn(MapboxExperimental::class)
   @RenderThread
   override fun doFrame(frameTimeNanos: Long) {
     trace("do-frame") {
-      val startTime = if (renderThreadRecorder?.recording == true) {
+      val startTime = if (renderThreadStatsRecorder?.isRecording == true) {
         SystemClock.elapsedRealtimeNanos()
       } else {
         0L
@@ -620,13 +623,13 @@ internal class MapboxRenderThread : Choreographer.FrameCallback {
       // With `awaitingNextVsync = false` we will always schedule recursive tasks for later execution
       // via `renderHandlerThread.postDelayed` instead of updating queue concurrently that is being drained (which may lead to deadlock in core).
       drainQueue(nonRenderEventQueue)
-      val endTime = if (renderThreadRecorder?.recording == true) {
+      val endTime = if (renderThreadStatsRecorder?.isRecording == true) {
         SystemClock.elapsedRealtimeNanos()
       } else {
         0L
       }
       if (startTime != 0L && endTime != 0L) {
-        renderThreadRecorder?.addFrameStats(
+        renderThreadStatsRecorder?.addFrameStats(
           (endTime - startTime) / 1e6,
           fpsManager.skippedNow
         )